# PS-3702: LP#1690588: ALTER TABLE is slow in 5.7 when the number of di…